server/pyproject.toml
[tool.poetry]
name = "yamt"
version = "0.1.0"
description = ""
authors = ["vodkar <boombarah@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.8.2"
pysnmp = "^4.4.12"
asyncssh = "^2.7.2"
scapy = "^2.4.5"
prompt-toolkit = "^3.0.23"
aioreactive = "^0.15.0"
aiofile = "^3.7.4"
PyYAML = "^6.0"
cryptography = "36.0.2"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
types-PyYAML = "^6.0.8"
pytest = "^6.2.5"
black = "^22.3"
pytest-asyncio = "^0.16.0"
isort = "^5.10.1"
mypy = "^0.910"
Faker = "^13.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 109

[tool.isort]
profile = "black"
